k8sagent support for apiextensions.k8s.io/v1 crd to support k8s 1.22+
Testing
 `agent-operator git:(agent-crdv1-changes) ✗ kubectl version
WARNING: This version information is deprecated and will be replaced with the output from kubectl version --short.  Use --output=yaml|json to get the full version.
Client Version: version.Info{Major:"1", Minor:"24", GitVersion:"v1.24.2", GitCommit:"f66044f4361b9f1f96f0053dd46cb7dce5e990a8", GitTreeState:"clean", BuildDate:"2022-06-15T14:14:10Z", GoVersion:"go1.18.3", Compiler:"gc", Platform:"darwin/arm64"}
Kustomize Version: v4.5.4
Server Version: version.Info{Major:"1", Minor:"24", GitVersion:"v1.24.1", GitCommit:"3ddd0f45aa91e2f30c70734b175631bec5b5825a", GitTreeState:"clean", BuildDate:"2022-05-24T12:18:48Z", GoVersion:"go1.18.2", Compiler:"gc", Platform:"linux/arm64"}
`


` agent-operator git:(agent-crdv1-changes) ✗ minikube kubectl -- logs -f concord-agent-operator-55d694858d-8c2d4
05:24:04.195 [main] [INFO ] c.w.concord.agentoperator.Operator - main -> my watch begins... (namespace=default)
05:24:36.597 [scheduler-worker] [INFO ] c.w.c.a.planner.CreatePodChange - apply -> created a pod example-agentpool/example-agentpool-00000
05:25:48.015 [scheduler-worker] [INFO ] c.w.c.a.p.CreateConfigMapChange - apply -> created a configmap example-agentpool-cfg
05:25:51.306 [scheduler-worker] [INFO ] c.w.concord.agentoperator.PodUtils - ['example-agentpool-00000']: apply tag ['concordTaggedForRemoval': 'true'] -> done
05:25:58.204 [scheduler-worker] [INFO ] c.w.concord.agentoperator.PodUtils - ['example-agentpool-00000']: apply tag ['preStopHookTermination': 'true'] -> done
05:25:58.204 [scheduler-worker] [INFO ] c.w.c.a.planner.TryToDeletePodChange - apply ['example-agentpool-00000'] -> Marked for termination (former phase: Running)
05:26:04.304 [scheduler-worker] [WARN ] c.w.c.a.planner.TryToDeletePodChange - ['example-agentpool-00000'] -> has been marked for termination. Waiting for running process to complete ...
05:26:09.802 [scheduler-worker] [WARN ] c.w.c.a.planner.TryToDeletePodChange - ['example-agentpool-00000'] -> has been marked for termination. Waiting for running process to complete ...
05:26:15.387 [scheduler-worker] [WARN ] c.w.c.a.planner.TryToDeletePodChange - ['example-agentpool-00000'] -> has been marked for termination. Waiting for running process to complete ...
05:26:21.086 [scheduler-worker] [WARN ] c.w.c.a.planner.TryToDeletePodChange - ['example-agentpool-00000'] -> has been marked for termination. Waiting for running process to complete ...
05:26:26.793 [scheduler-worker] [INFO ] c.w.c.a.planner.CreatePodChange - apply -> created a pod example-agentpool/example-agentpool-00000
05:26:40.698 [scheduler-worker] [INFO ] c.w.c.a.planner.CreatePodChange - apply -> created a pod example-agentpool/example-agentpool-00000
`

`➜  agent-operator git:(agent-crdv1-changes) ✗ kubectl get crds
NAME                                 CREATED AT
agentpools.concord.walmartlabs.com   2022-07-14T05:18:13Z
➜  agent-operator git:(agent-crdv1-changes) ✗ kubectl get agentpools
NAME                AGE
example-agentpool   29m
➜  agent-operator git:(agent-crdv1-changes) ✗ kubectl get agentpool/example-agentpool
NAME                AGE
example-agentpool   29m
`
