+++ k8s/agent-operator/src/main/java/com/walmartlabs/concord/agentoperator/crd/AgentPoolList.java
@@ -20,9 +20,12 @@ package com.walmartlabs.concord.agentoperator.crd;
  * =====
  */
 
+import io.fabric8.kubernetes.api.model.KubernetesResourceList;
+import io.fabric8.kubernetes.api.model.ListMeta;
 import io.fabric8.kubernetes.client.CustomResourceList;
+
+import java.util.List;
 
 public class AgentPoolList extends CustomResourceList<AgentPool> {
 
-    private static final long serialVersionUID = 1L;
 }
