--- k8s/agent-operator/src/main/java/com/walmartlabs/concord/agentoperator/crd/DoneableAgentPool.java
@@ -1,31 +1,0 @@
-package com.walmartlabs.concord.agentoperator.crd;
-
-/*-
- * *****
- * Concord
- * -----
- * Copyright (C) 2017 - 2019 Walmart Inc.
- * -----
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * =====
- */
-
-import io.fabric8.kubernetes.api.builder.Function;
-import io.fabric8.kubernetes.client.CustomResourceDoneable;
-
-public class DoneableAgentPool extends CustomResourceDoneable<AgentPool> {
-
-    public DoneableAgentPool(AgentPool resource, Function<AgentPool, AgentPool> function) {
-        super(resource, function);
-    }
-}
