--- k8s/agent-operator/src/main/resources/META-INF/services/io.fabric8.kubernetes.api.KubernetesResourceMappingProvider
@@ -1,2 +1,0 @@
-com.walmartlabs.concord.agentoperator.crd.AgentPoolMappingProvider
-io.fabric8.kubernetes.internal.InternalResourceMappingProvider
