+++ pom.xml
@@ -205,7 +205,7 @@
                 <plugin>
                     <groupId>io.takari.maven.plugins</groupId>
                     <artifactId>takari-lifecycle-plugin</artifactId>
-                    <version>1.13.7</version>
+                    <version>2.0.8</version>
                     <extensions>true</extensions>
                     <configuration>
                         <source>${maven.compiler.source}</source>
